+++ apps/prairielearn/src/pages/instructorGradingJob/instructorGradingJob.sql
@@ -44,15 +44,16 @@ FROM
   grading_jobs AS gj
   JOIN submissions AS s ON (s.id = gj.submission_id)
   JOIN variants AS v ON (v.id = s.variant_id)
+  JOIN pl_courses AS c ON (c.id = v.course_id)
   JOIN questions AS q ON (q.id = v.question_id)
   LEFT JOIN instance_questions AS iq ON (iq.id = v.instance_question_id)
   LEFT JOIN assessment_questions AS aq ON (aq.id = iq.assessment_question_id)
   LEFT JOIN assessment_instances AS ai ON (ai.id = iq.assessment_instance_id)
   LEFT JOIN assessments AS a ON (a.id = ai.assessment_id)
-  LEFT JOIN pl_courses AS c ON (c.id = v.course_id)
-  LEFT JOIN course_instances AS ci ON (ci.course_id = c.id)
+  LEFT JOIN course_instances AS ci ON (ci.id = v.course_instance_id)
   LEFT JOIN users AS u ON (u.user_id = s.auth_user_id)
 WHERE
   gj.id = $job_id
   AND gj.grading_method = 'External'
-  AND ci.id = $course_instance_id;
+  AND c.id = $course_id
+  AND ci.id IS NOT DISTINCT FROM $course_instance_id;