+++ apps/prairielearn/src/tests/manualGrading.test.ts
@@ -30,6 +30,7 @@ interface RubricItem {
   description: string;
   explanation?: string;
   grader_note?: string;
+  always_show_to_students: boolean;
   description_render?: string;
   explanation_render?: string;
   grader_note_render?: string;
@@ -226,26 +227,30 @@ function checkGradingResults(assigned_grader: MockUser, grader: MockUser): void 
     } else {
       rubric_items.forEach((item, index) => {
         const container = feedbackBlock.find(`[data-testid="rubric-item-container-${item.id}"]`);
-        assert.equal(container.length, 1);
-        assert.equal(
-          container.find('input[type="checkbox"]').is(':checked'),
-          selected_rubric_items.includes(index),
-        );
-        assert.equal(
-          container.find('[data-testid="rubric-item-points"]').text().trim(),
-          `[${item.points >= 0 ? '+' : ''}${item.points}]`,
-        );
-        assert.equal(
-          container.find('[data-testid="rubric-item-description"]').html()?.trim(),
-          item.description_render ?? item.description,
-        );
-        if (item.explanation) {
+        if (item.always_show_to_students || selected_rubric_items.includes(index)) {
+          assert.equal(container.length, 1);
           assert.equal(
-            container.find('[data-testid="rubric-item-explanation"]').attr('data-content'),
-            item.explanation_render ?? `<p>${item.explanation}</p>`,
+            container.find('input[type="checkbox"]').is(':checked'),
+            selected_rubric_items.includes(index),
           );
+          assert.equal(
+            container.find('[data-testid="rubric-item-points"]').text().trim(),
+            `[${item.points >= 0 ? '+' : ''}${item.points}]`,
+          );
+          assert.equal(
+            container.find('[data-testid="rubric-item-description"]').html()?.trim(),
+            item.description_render ?? item.description,
+          );
+          if (item.explanation) {
+            assert.equal(
+              container.find('[data-testid="rubric-item-explanation"]').attr('data-content'),
+              item.explanation_render ?? `<p>${item.explanation}</p>`,
+            );
+          } else {
+            assert.equal(container.find('[data-testid="rubric-item-explanation"]').length, 0);
+          }
         } else {
-          assert.equal(container.find('[data-testid="rubric-item-explanation"]').length, 0);
+          assert.equal(container.length, 0);
         }
       });
     }
@@ -299,6 +304,10 @@ function checkSettingsResults(
         `label[data-input-name="rubric_item[cur${item.id}][grader_note]"]`,
       );
       assert.equal(graderNote.attr('data-current-value') ?? '', item.grader_note ?? '');
+      const always_show_to_students = form.find(
+        `[name="rubric_item[cur${item.id}][always_show_to_students]"]:checked`,
+      );
+      assert.equal(always_show_to_students.val(), item.always_show_to_students ? 'true' : 'false');
     });
   });
 
@@ -690,7 +699,7 @@ describe('Manual Grading', function () {
           const $manualGradingIQPage = cheerio.load(manualGradingIQPage);
           const form = $manualGradingIQPage('form[name=rubric-settings]');
           rubric_items = [
-            { points: 6, description: 'First rubric item' },
+            { points: 6, description: 'First rubric item', always_show_to_students: true },
             {
               points: 3,
               description: 'Second rubric item (partial, with `markdown`)',
@@ -699,6 +708,7 @@ describe('Manual Grading', function () {
               description_render: 'Second rubric item (partial, with <code>markdown</code>)',
               explanation_render: '<p>Explanation with <strong>markdown</strong></p>',
               grader_note_render: '<p>Instructions with <em>markdown</em></p>',
+              always_show_to_students: false,
             },
             {
               points: 0.4,
@@ -710,14 +720,17 @@ describe('Manual Grading', function () {
               explanation_render: '<p>Explanation with moustache: 43</p>',
               grader_note_render:
                 '<p>Instructions with <em>markdown</em> and moustache: 49</p>\n<p>And more than one line</p>',
+              always_show_to_students: true,
             },
             {
               points: -1.6,
               description: 'Penalty rubric item (negative points, floating point)',
+              always_show_to_students: false,
             },
             {
               points: 0,
               description: 'Rubric item with no value (zero points)',
+              always_show_to_students: true,
             },
           ];
 
@@ -762,6 +775,46 @@ describe('Manual Grading', function () {
           rubric_items[2].points = 1;
           score_points = 5.4;
           score_percent = 90;
+
+          const response = await fetch(manualGradingIQUrl, {
+            method: 'POST',
+            headers: { 'Content-type': 'application/x-www-form-urlencoded' },
+            body: new URLSearchParams({
+              __action: form.find('input[name=__action]').attr('value') || '',
+              __csrf_token: form.find('input[name=__csrf_token]').attr('value') || '',
+              modified_at: form.find('input[name=modified_at]').attr('value') || '',
+              use_rubric: 'true',
+              starting_points: '0', // Positive grading
+              min_points: '-0.5',
+              max_extra_points: '0.5',
+              ...buildRubricItemFields(rubric_items),
+            }).toString(),
+          });
+
+          assert.equal(response.ok, true);
+        });
+
+        checkSettingsResults(0, -0.5, 0.5);
+        checkGradingResults(mockStaff[0], mockStaff[0]);
+      });
+
+      describe('Grading without rubric items', () => {
+        step('submit a grade using a positive rubric', async () => {
+          setUser(mockStaff[0]);
+          selected_rubric_items = [];
+          score_points = 0;
+          score_percent = 0;
+          feedback_note = 'Test feedback note without any rubric items';
+          await submitGradeForm();
+        });
+
+        checkGradingResults(mockStaff[0], mockStaff[0]);
+
+        step('update rubric items should succeed', async () => {
+          setUser(mockStaff[0]);
+          const manualGradingIQPage = await (await fetch(manualGradingIQUrl)).text();
+          const $manualGradingIQPage = cheerio.load(manualGradingIQPage);
+          const form = $manualGradingIQPage('form[name=rubric-settings]');
 
           const response = await fetch(manualGradingIQUrl, {
             method: 'POST',
@@ -873,7 +926,7 @@ describe('Manual Grading', function () {
           const $manualGradingIQPage = cheerio.load(manualGradingIQPage);
           const form = $manualGradingIQPage('form[name=rubric-settings]');
           rubric_items = [
-            { points: 0, description: 'First rubric item' },
+            { points: 0, description: 'First rubric item', always_show_to_students: true },
             {
               points: -3,
               description: 'Second rubric item (partial, with `markdown`)',
@@ -882,6 +935,7 @@ describe('Manual Grading', function () {
               description_render: 'Second rubric item (partial, with <code>markdown</code>)',
               explanation_render: '<p>Explanation with <strong>markdown</strong></p>',
               grader_note_render: '<p>Instructions with <em>markdown</em></p>',
+              always_show_to_students: true,
             },
             {
               points: -4,
@@ -893,15 +947,18 @@ describe('Manual Grading', function () {
               explanation_render: '<p>Explanation with moustache: 43</p>',
               grader_note_render:
                 '<p>Instructions with <em>markdown</em> and moustache: 49</p>\n<p>And more than one line</p>',
+              always_show_to_students: false,
             },
             {
               points: 1.6,
               description:
                 'Positive rubric item in negative grading (positive points, floating point)',
+              always_show_to_students: false,
             },
             {
               points: 6,
               description: 'Rubric item with positive reaching maximum',
+              always_show_to_students: true,
             },
           ];