+++ packages/flash/src/index.test.ts
@@ -1,6 +1,7 @@
 import { assert } from 'chai';
 
 import { flashMiddleware, flash } from './index';
+import { html } from '@prairielearn/html';
 
 describe('flash', () => {
   it('throws an error if no session present', () => {
@@ -13,9 +14,12 @@ describe('flash', () => {
     assert.throw(() => {
       flash('notice', 'Hello world');
     });
+    assert.throw(() => {
+      flash('notice', html`<p>hello ${'&'} world</p>`);
+    });
   });
 
-  it('adds a flash', () => {
+  it('adds a flash using string message', () => {
     const req = {
       session: {},
     } as any;
@@ -28,6 +32,34 @@ describe('flash', () => {
     });
   });
 
+  it('adds a flash and escapes unsafe HTML', () => {
+    const req = {
+      session: {},
+    } as any;
+    const res = {} as any;
+
+    flashMiddleware()(req, res, () => {
+      flash('notice', '<b>hello world</b>');
+
+      assert.sameDeepMembers(flash(), [
+        { type: 'notice', message: '&lt;b&gt;hello world&lt;/b&gt;' },
+      ]);
+    });
+  });
+
+  it('adds a flash with HTML-safe message', () => {
+    const req = {
+      session: {},
+    } as any;
+    const res = {} as any;
+
+    flashMiddleware()(req, res, () => {
+      flash('notice', html`<p>hello ${'&'} world</p>`);
+
+      assert.sameDeepMembers(flash(), [{ type: 'notice', message: '<p>hello &amp; world</p>' }]);
+    });
+  });
+
   it('stores multiples flashes with the same type', () => {
     const req = {
       session: {},
@@ -36,11 +68,13 @@ describe('flash', () => {
 
     flashMiddleware()(req, res, () => {
       flash('notice', 'hello world');
-      flash('notice', 'goodbye world');
+      flash('notice', '<< goodbye world');
+      flash('notice', html`<p>hello ${'&'} world</p>`);
 
       assert.sameDeepMembers(flash(), [
         { type: 'notice', message: 'hello world' },
-        { type: 'notice', message: 'goodbye world' },
+        { type: 'notice', message: '&lt;&lt; goodbye world' },
+        { type: 'notice', message: '<p>hello &amp; world</p>' },
       ]);
     });
   });
@@ -53,10 +87,16 @@ describe('flash', () => {
 
     flashMiddleware()(req, res, () => {
       flash('notice', 'hello world');
-      flash('error', 'goodbye world');
+      flash('error', '<< goodbye world');
+      flash('success', html`<p>hello ${'&'} world</p>`);
 
       assert.sameDeepMembers(flash('notice'), [{ type: 'notice', message: 'hello world' }]);
-      assert.sameDeepMembers(flash('error'), [{ type: 'error', message: 'goodbye world' }]);
+      assert.sameDeepMembers(flash('error'), [
+        { type: 'error', message: '&lt;&lt; goodbye world' },
+      ]);
+      assert.sameDeepMembers(flash('success'), [
+        { type: 'success', message: '<p>hello &amp; world</p>' },
+      ]);
     });
   });
 
@@ -68,11 +108,13 @@ describe('flash', () => {
 
     flashMiddleware()(req, res, () => {
       flash('notice', 'hello world');
-      flash('error', 'goodbye world');
+      flash('error', '<< goodbye world');
+      flash('success', html`<p>hello ${'&'} world</p>`);
 
       assert.sameDeepMembers(flash(), [
         { type: 'notice', message: 'hello world' },
-        { type: 'error', message: 'goodbye world' },
+        { type: 'error', message: '&lt;&lt; goodbye world' },
+        { type: 'success', message: '<p>hello &amp; world</p>' },
       ]);
     });
   });
@@ -85,13 +127,20 @@ describe('flash', () => {
 
     flashMiddleware()(req, res, () => {
       flash('notice', 'hello world');
-      flash('error', 'goodbye world');
+      flash('error', '<< goodbye world');
+      flash('success', html`<p>hello ${'&'} world</p>`);
 
       assert.sameDeepMembers(flash('notice'), [{ type: 'notice', message: 'hello world' }]);
-      assert.sameDeepMembers(flash('error'), [{ type: 'error', message: 'goodbye world' }]);
+      assert.sameDeepMembers(flash('error'), [
+        { type: 'error', message: '&lt;&lt; goodbye world' },
+      ]);
+      assert.sameDeepMembers(flash('success'), [
+        { type: 'success', message: '<p>hello &amp; world</p>' },
+      ]);
 
       assert.deepEqual(flash('notice'), []);
       assert.deepEqual(flash('error'), []);
+      assert.deepEqual(flash('success'), []);
       assert.isEmpty(flash());
     });
   });