+++ tools/build-image-if-needed.sh
@@ -25,7 +25,6 @@ if git diff --exit-code $DIFF_SOURCE...HEAD -- ${BUILD_DIRECTORY}; then
   echo "${BUILD_DIRECTORY} files not modified; no rebuild required"
 else
   echo "${BUILD_DIRECTORY} files modified; ${TAG_NAME} requires a rebuild"
-  docker buildx build --load ${BUILD_DIRECTORY} -t ${TAG_NAME}
-  docker image history ${TAG_NAME}
+  docker buildx build ${BUILD_DIRECTORY} -t ${TAG_NAME}
 fi