+++ packages/eslint-plugin-prairielearn/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/eslint-plugin
 
+## 1.0.3
+
+### Patch Changes
+
+- 1523b97b0: Upgrade all dependencies
+
+## 1.0.2
+
+### Patch Changes
+
+- 8dd894623: Upgrade all dependencies
+
 ## 1.0.1
 
 ### Patch Changes