+++ packages/prettier-plugin-sql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/prettier-plugin-sql
 
+## 1.0.9
+
+### Patch Changes
+
+- 1523b97b0: Upgrade all dependencies
+
+## 1.0.8
+
+### Patch Changes
+
+- 8dd894623: Upgrade all dependencies
+
 ## 1.0.7
 
 ### Patch Changes